kernel: machp()->ticks and sys->ticks become 64bit
The ticks semantics do not change: they increment with hzclock runs on
cpu0 (see apiconline on 9/k10/apic.c)
Given that nsec(2) is based on the RDTSC instruction, we can use
one clock to check the other.

regress/alarm		PASS
regress/nanotime	PASS
regress/tsemacquire	PASS

+++ sys/src/9/port/devaoe.c
@@ -161,7 +161,7 @@ typedef struct Srb Srb;
 struct Srb {
 	Rendez;
 	Srb	*next;
-	uint32_t	ticksent;
+	uint64_t	ticksent;
 	uint32_t	len;
 	int64_t	sector;
 	short	write;
@@ -176,7 +176,7 @@ typedef struct {
 	uint32_t	bcnt;
 	uint32_t	dlen;
 	int64_t	lba;
-	uint32_t	ticksent;
+	uint64_t	ticksent;
 	int	nhdr;
 	uint8_t	hdr[ETHERMINTU];
 	void	*dp;
@@ -214,7 +214,7 @@ struct Aoedev {
 	uint32_t	lostjumbo;
 	uint16_t	nout;
 	uint16_t	maxout;
-	uint32_t	lastwadj;
+	uint64_t	lastwadj;
 	Srb	*_head;
 	Srb	*_tail;
 	Srb	*inprocess;
